+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Add `CheckBytes` impl for `rkyv`ed structs
 - Add `rkyv` implementations behind feature [#69](https://github.com/dusk-network/schnorr/issues/69)
 
 ### Changed
@@ -16,7 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Update dusk-poseidon from `0.23.0-rc` to `0.26`
 - Update dusk-pki from `0.9.0-rc` to `0.11`
 - Update dusk-plonk from `0.9` to `0.12`
-- Update dusk-bls12_381 from `0.8` to `0.10`
+- Update dusk-bls12_381 from `0.8` to `0.11`
 - Update dusk-jubjub from `0.10` to `0.12`
 - Update canonical from `0.6` to `0.7`
 - Update canonical_derive from `0.6` to `0.7`